+++ src/main/java/app/DAO/DAOinterfaces/ControllersDAO.java
@@ -7,10 +7,10 @@ import java.util.Collection;
 public interface ControllersDAO extends DAOMain<ControllersEntity> {
 
     public Collection<ControllersEntity> getUsers(int limit, int offset, ControllersEntity controllersEntity);
-    public ControllersEntity getControllerModelByModel(int idController);
-    public ControllersEntity getWorkStatusByWorkStatus(int idController);
-    public ControllersEntity getDevicesByIdController(int idController);
-    public ControllersEntity getGroupsControllersByIdController(int idController);
-    public ControllersEntity getControllerById(int idController);
+    public ControllersEntity getControllerModelByModel(Long idController);
+    public ControllersEntity getWorkStatusByWorkStatus(Long idController);
+    public ControllersEntity getDevicesByIdController(Long idController);
+    public ControllersEntity getGroupsControllersByIdController(Long idController);
+    public ControllersEntity getControllerById(Long idController);
 
 }
